Refactor CLI commands use data transfer objects Don't support multiple user types
